Chapter 12
These are the challenges from Chapter 12.
The first one creates a custom function called countstrings that
determines the number of fields of type in string in an input feature
class. The second challenge modifies the parcel.tax file so that it
determines the property tax for each parcel and stores the values in a
list.