Chapter 13
These are the challenges from chapter 13. This script makes a copy of the
random percent file and modifies the script to add a third parameter that
is a percentage of the number of input records as an integer between one
and 100. The tool is validated on the tool dialog box.